## Proposed Changes

Serializing the `decimal` datatype allocates 40 bytes every time and deserializing it is sub-optimal. This optimization removes that allocation and speeds up (de)serializing decimal values

## Types of Changes

What types of changes does your code introduce to this project?
_Put an `x` in the boxes that apply_

- [ ] Bug fix (non-breaking change which fixes issue #NNNN)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause an observable behavior change in existing systems)
- [ ] Documentation improvements (corrections, new content, etc)
- [ ] Cosmetic change (whitespace, formatting, etc)

## Checklist

- [X] I have read the `CONTRIBUTING.md` document
- [X] I have signed the CA (see https://cla.pivotal.io/sign/rabbitmq)
- [X] All tests pass locally with my changes
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)
- [ ] Any dependent changes have been merged and published in related repositories

## Results
``` ini
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.20270
Intel Core i7-10700 CPU 2.90GHz, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=5.0.101
  [Host]        : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  .NET 4.8      : .NET Framework 4.8 (4.8.4261.0), X64 RyuJIT
  .NET Core 3.1 : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
```
### Before
|         Method |           Job |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|--------------- |-------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|----------:|
|    DecimalRead |      .NET 4.8 | 21.939 ns | 0.1721 ns | 0.1526 ns |      - |     - |     - |         - |     781 B |
|   DecimalWrite |      .NET 4.8 | 29.016 ns | 0.0880 ns | 0.0735 ns | 0.0063 |     - |     - |      40 B |     686 B |
|    DecimalRead | .NET Core 3.1 |  5.785 ns | 0.0151 ns | 0.0133 ns |      - |     - |     - |         - |     449 B |
|   DecimalWrite | .NET Core 3.1 | 14.726 ns | 0.2515 ns | 0.2353 ns | 0.0048 |     - |     - |      40 B |     300 B |

### After
|       Method |           Job |      Mean |     Error |    StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated | Code Size |
|------------- |-------------- |----------:|----------:|----------:|------:|------:|------:|----------:|----------:|
|  DecimalRead |      .NET 4.8 | 12.192 ns | 0.0406 ns | 0.0360 ns |     - |     - |     - |         - |     592 B |
| DecimalWrite |      .NET 4.8 | 13.177 ns | 0.1229 ns | 0.1089 ns |     - |     - |     - |         - |     659 B |
|  DecimalRead | .NET Core 3.1 |  2.439 ns | 0.0136 ns | 0.0121 ns |     - |     - |     - |         - |     227 B |
| DecimalWrite | .NET Core 3.1 |  3.068 ns | 0.0142 ns | 0.0126 ns |     - |     - |     - |         - |     284 B |