@bollhals 2020-12-09
I assume decimals aren't used as often, but the improvement is substantial nonetheless. 👍