@stebet 2020-12-09
> I assume decimals aren't used as often, but the improvement is substantial nonetheless. 👍

I figured since it's a user-provided field, it might be used considerably in some cases (financial apps come to mind), so better to have most primitives as fast as possible.