@stebet 2020-12-16
> I assume this can be backported all the way to `6.x`?

Yup :)